requirements.txt
# python-telegram-bot==12.7
requests==2.31.0
pyfiglet==0.8.post1
python-dotenv===0.17.1

psycopg2==2.9.9
redshift_connector==2.0.917
awswrangler==3.4.2
PyYAML==6.0.2
openai==0.27.8
boto3==1.28.3
botocore==1.31.3
pandas==2.1.3
flask==3.0.3
bs4==0.0.2
beautifulsoup4==4.12.2
python-telegram-bot==21.6
pyaml==24.9.0